+++ db/db_filesnapshot.cc
@@ -119,7 +119,7 @@ Status DBImpl::GetLiveFiles(std::vector<std::string>& ret,
   }
 
   // Make a set of all of the live *.sst files
-  std::vector<FileDescriptor> live;
+  std::vector<uint64_t> live;
   for (auto cfd : *versions_->GetColumnFamilySet()) {
     if (cfd->IsDropped()) {
       continue;
@@ -133,7 +133,7 @@ Status DBImpl::GetLiveFiles(std::vector<std::string>& ret,
   // create names of the live files. The names are not absolute
   // paths, instead they are relative to dbname_;
   for (const auto& live_file : live) {
-    ret.push_back(MakeTableFileName("", live_file.GetNumber()));
+    ret.push_back(MakeTableFileName("", live_file));
   }
 
   ret.push_back(CurrentFileName(""));