+++ db/version_set.cc
@@ -742,9 +742,7 @@ Version::~Version() {
   for (int level = 0; level < storage_info_.num_levels_; level++) {
     for (size_t i = 0; i < storage_info_.files_[level].size(); i++) {
       FileMetaData* f = storage_info_.files_[level][i];
-      assert(f->refs > 0);
-      f->refs--;
-      if (f->refs <= 0) {
+      if (f->Unref()) {
         assert(cfd_ != nullptr);
         uint32_t path_id = f->fd.GetPathId();
         assert(path_id < cfd_->ioptions()->cf_paths.size());
@@ -2297,6 +2295,7 @@ void VersionStorageInfo::EstimateCompactionBytesNeeded(
 namespace {
 uint32_t GetExpiredTtlFilesCount(const ImmutableCFOptions& ioptions,
                                  const MutableCFOptions& mutable_cf_options,
+                                 int level,
                                  const std::vector<FileMetaData*>& files) {
   uint32_t ttl_expired_files_count = 0;
 
@@ -2305,7 +2304,8 @@ uint32_t GetExpiredTtlFilesCount(const ImmutableCFOptions& ioptions,
   if (status.ok()) {
     const uint64_t current_time = static_cast<uint64_t>(_current_time);
     for (auto f : files) {
-      if (!f->being_compacted && f->fd.table_reader != nullptr &&
+      if ((!f->being_compacted || f->being_moved_to == level) &&
+          f->fd.table_reader != nullptr &&
           f->fd.table_reader->GetTableProperties() != nullptr) {
         auto creation_time =
             f->fd.table_reader->GetTableProperties()->creation_time;
@@ -2340,7 +2340,7 @@ void VersionStorageInfo::ComputeCompactionScore(
       int num_sorted_runs = 0;
       uint64_t total_size = 0;
       for (auto* f : files_[level]) {
-        if (!f->being_compacted) {
+        if (!f->being_compacted || f->being_moved_to == level) {
           total_size += f->compensated_file_size;
           num_sorted_runs++;
         }
@@ -2350,7 +2350,8 @@ void VersionStorageInfo::ComputeCompactionScore(
         // compaction score for the whole DB. Adding other levels as if
         // they are L0 files.
         for (int i = 1; i < num_levels(); i++) {
-          if (!files_[i].empty() && !files_[i][0]->being_compacted) {
+          if (!files_[i].empty() && (!files_[i][0]->being_compacted ||
+                                     files_[i][0]->being_moved_to == i)) {
             num_sorted_runs++;
           }
         }
@@ -2366,10 +2367,10 @@ void VersionStorageInfo::ComputeCompactionScore(
               score);
         }
         if (mutable_cf_options.ttl > 0) {
-          score = std::max(
-              static_cast<double>(GetExpiredTtlFilesCount(
-                  immutable_cf_options, mutable_cf_options, files_[level])),
-              score);
+          score = std::max(static_cast<double>(GetExpiredTtlFilesCount(
+                               immutable_cf_options, mutable_cf_options, level,
+                               files_[level])),
+                           score);
         }
 
       } else {
@@ -2388,7 +2389,7 @@ void VersionStorageInfo::ComputeCompactionScore(
       // Compute the ratio of current size to size limit.
       uint64_t level_bytes_no_compacting = 0;
       for (auto f : files_[level]) {
-        if (!f->being_compacted) {
+        if (!f->being_compacted || f->being_moved_to == level) {
           level_bytes_no_compacting += f->compensated_file_size;
         }
       }
@@ -2441,7 +2442,8 @@ void VersionStorageInfo::ComputeFilesMarkedForCompaction() {
 
   for (int level = 0; level <= last_qualify_level; level++) {
     for (auto* f : files_[level]) {
-      if (!f->being_compacted && f->marked_for_compaction) {
+      if ((!f->being_compacted || f->being_moved_to == level) &&
+          f->marked_for_compaction) {
         files_marked_for_compaction_.emplace_back(level, f);
       }
     }
@@ -2463,7 +2465,8 @@ void VersionStorageInfo::ComputeExpiredTtlFiles(
 
   for (int level = 0; level < num_levels() - 1; level++) {
     for (auto f : files_[level]) {
-      if (!f->being_compacted && f->fd.table_reader != nullptr &&
+      if ((!f->being_compacted || f->being_moved_to == level) &&
+          f->fd.table_reader != nullptr &&
           f->fd.table_reader->GetTableProperties() != nullptr) {
         auto creation_time =
             f->fd.table_reader->GetTableProperties()->creation_time;
@@ -2493,7 +2496,8 @@ void VersionStorageInfo::ComputeFilesMarkedForPeriodicCompaction(
 
   for (int level = 0; level < num_levels(); level++) {
     for (auto f : files_[level]) {
-      if (!f->being_compacted && f->fd.table_reader != nullptr &&
+      if ((!f->being_compacted || f->being_moved_to == level) &&
+          f->fd.table_reader != nullptr &&
           f->fd.table_reader->GetTableProperties() != nullptr) {
         // Compute a file's modification time in the following order:
         // 1. Use file_creation_time table property if it is > 0.
@@ -2545,13 +2549,13 @@ bool CompareCompensatedSizeDescending(const Fsize& first, const Fsize& second) {
 } // anonymous namespace
 
 void VersionStorageInfo::AddFile(int level, FileMetaData* f, Logger* info_log) {
-  auto* level_files = &files_[level];
+  auto& level_files = files_[level];
   // Must not overlap
 #ifndef NDEBUG
-  if (level > 0 && !level_files->empty() &&
+  if (level > 0 && !level_files.empty() &&
       internal_comparator_->Compare(
-          (*level_files)[level_files->size() - 1]->largest, f->smallest) >= 0) {
-    auto* f2 = (*level_files)[level_files->size() - 1];
+          level_files[level_files.size() - 1]->largest, f->smallest) >= 0) {
+    auto* f2 = level_files[level_files.size() - 1];
     if (info_log != nullptr) {
       Error(info_log, "Adding new file %" PRIu64
                       " range (%s, %s) to level %d but overlapping "
@@ -2567,14 +2571,15 @@ void VersionStorageInfo::AddFile(int level, FileMetaData* f, Logger* info_log) {
 #else
   (void)info_log;
 #endif
-  f->refs++;
-  level_files->push_back(f);
+  level_files.push_back(f);
+
+  f->Ref();
 
   const uint64_t file_number = f->fd.GetNumber();
 
   assert(file_locations_.find(file_number) == file_locations_.end());
   file_locations_.emplace(file_number,
-                          FileLocation(level, level_files->size() - 1));
+                          FileLocation(level, level_files.size() - 1));
 }
 
 // Version::PrepareApply() need to be called before calling the function, or
@@ -2616,6 +2621,11 @@ void VersionStorageInfo::SetFinalized() {
     }
     if (LevelFiles(level).size() > 0) {
       assert(level < num_non_empty_levels());
+    }
+    for (auto* f : LevelFiles(level)) {
+      if (f->being_moved_to == level) {
+        f->being_moved_to = -1;
+      }
     }
   }
   assert(compaction_level_.size() > 0);
@@ -3366,11 +3376,13 @@ bool VersionStorageInfo::RangeMightExistAfterSortedRun(
   return false;
 }
 
-void Version::AddLiveFiles(std::vector<FileDescriptor>* live) {
-  for (int level = 0; level < storage_info_.num_levels(); level++) {
-    const std::vector<FileMetaData*>& files = storage_info_.files_[level];
-    for (const auto& file : files) {
-      live->push_back(file->fd);
+void Version::AddLiveFiles(std::vector<uint64_t>* live_table_files) const {
+  assert(live_table_files);
+  for (int level = 0; level < storage_info_.num_levels(); ++level) {
+    const auto& level_files = storage_info_.LevelFiles(level);
+    for (const auto& meta : level_files) {
+      assert(meta);
+      live_table_files->emplace_back(meta->fd.GetNumber());
     }
   }
 }
@@ -5119,9 +5131,10 @@ uint64_t VersionSet::ApproximateSize(Version* v, const FdWithKeyRange& f,
   return result;
 }
 
-void VersionSet::AddLiveFiles(std::vector<FileDescriptor>* live_list) {
+void VersionSet::AddLiveFiles(std::vector<uint64_t>* live_table_files) const {
+  assert(live_table_files);
   // pre-calculate space requirement
-  int64_t total_files = 0;
+  int64_t total_table_files = 0;
   for (auto cfd : *column_family_set_) {
     if (!cfd->initialized()) {
       continue;
@@ -5131,13 +5144,13 @@ void VersionSet::AddLiveFiles(std::vector<FileDescriptor>* live_list) {
          v = v->next_) {
       const auto* vstorage = v->storage_info();
       for (int level = 0; level < vstorage->num_levels(); level++) {
-        total_files += vstorage->LevelFiles(level).size();
+        total_table_files += vstorage->LevelFiles(level).size();
       }
     }
   }
 
   // just one time extension to the right size
-  live_list->reserve(live_list->size() + static_cast<size_t>(total_files));
+  live_table_files->reserve(live_table_files->size() + total_table_files);
 
   for (auto cfd : *column_family_set_) {
     if (!cfd->initialized()) {
@@ -5148,7 +5161,7 @@ void VersionSet::AddLiveFiles(std::vector<FileDescriptor>* live_list) {
     Version* dummy_versions = cfd->dummy_versions();
     for (Version* v = dummy_versions->next_; v != dummy_versions;
          v = v->next_) {
-      v->AddLiveFiles(live_list);
+      v->AddLiveFiles(live_table_files);
       if (v == current) {
         found_current = true;
       }
@@ -5156,7 +5169,7 @@ void VersionSet::AddLiveFiles(std::vector<FileDescriptor>* live_list) {
     if (!found_current && current != nullptr) {
       // Should never happen unless it is a bug.
       assert(false);
-      current->AddLiveFiles(live_list);
+      current->AddLiveFiles(live_table_files);
     }
   }
 }