+++ db/version_set_test.cc
@@ -118,7 +118,7 @@ class VersionStorageInfoTest : public testing::Test {
   ~VersionStorageInfoTest() override {
     for (int i = 0; i < vstorage_.num_levels(); i++) {
       for (auto* f : vstorage_.LevelFiles(i)) {
-        if (--f->refs == 0) {
+        if (f->Unref()) {
           delete f;
         }
       }
@@ -133,7 +133,6 @@ class VersionStorageInfoTest : public testing::Test {
     f->smallest = GetInternalKey(smallest, 0);
     f->largest = GetInternalKey(largest, 0);
     f->compensated_file_size = file_size;
-    f->refs = 0;
     f->num_entries = 0;
     f->num_deletions = 0;
     vstorage_.AddFile(level, f);
@@ -147,7 +146,6 @@ class VersionStorageInfoTest : public testing::Test {
     f->smallest = smallest;
     f->largest = largest;
     f->compensated_file_size = file_size;
-    f->refs = 0;
     f->num_entries = 0;
     f->num_deletions = 0;
     vstorage_.AddFile(level, f);
@@ -409,7 +407,7 @@ TEST_F(VersionStorageInfoTest, GetOverlappingInputs) {
       1, {"i", 0, kTypeValue}, {"j", 0, kTypeValue}));
 }
 
-TEST_F(VersionStorageInfoTest, FileLocation) {
+TEST_F(VersionStorageInfoTest, FileLocationAndMetaDataByNumber) {
   Add(0, 11U, "1", "2", 5000U);
   Add(0, 12U, "1", "2", 5000U);
 
@@ -417,13 +415,18 @@ TEST_F(VersionStorageInfoTest, FileLocation) {
 
   ASSERT_EQ(vstorage_.GetFileLocation(11U),
             VersionStorageInfo::FileLocation(0, 0));
+  ASSERT_NE(vstorage_.GetFileMetaDataByNumber(11U), nullptr);
+
   ASSERT_EQ(vstorage_.GetFileLocation(12U),
             VersionStorageInfo::FileLocation(0, 1));
+  ASSERT_NE(vstorage_.GetFileMetaDataByNumber(12U), nullptr);
 
   ASSERT_EQ(vstorage_.GetFileLocation(7U),
             VersionStorageInfo::FileLocation(2, 0));
+  ASSERT_NE(vstorage_.GetFileMetaDataByNumber(7U), nullptr);
 
   ASSERT_FALSE(vstorage_.GetFileLocation(999U).IsValid());
+  ASSERT_EQ(vstorage_.GetFileMetaDataByNumber(999U), nullptr);
 }
 
 class FindLevelFileTest : public testing::Test {